scraping-estimation-tool/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
configparser==5.0.0
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
Django==3.0.6
fake-useragent==0.1.11
fpdf==1.7.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
mysql-connector-python
mysqlclient==1.4.6
parse==1.15.0
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyee==7.0.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyppeteer==0.0.25
pyquery==1.4.1
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
requests-html==0.10.0
Scrapy==2.0.1
scrapy-user-agents==0.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stem==1.8.0
termcolor==1.1.0
tqdm==4.45.0
Twisted==20.3.0
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
w3lib==1.21.0
websockets==8.1
zope.interface==5.1.0
